requirements.txt
aldryn-boilerplates==0.7.5
aldryn-common==1.0.4
aldryn-search==0.3.0
appdirs==1.4.3
Babel==2.4.0
BabelDjango==0.2.2
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.4.17
chardet==3.0.3
cmsplugin-filer==0.10.2
contextlib2==0.5.5
dicttoxml==1.5.9
dj-database-url==0.4.2
Django==1.8.18
django-activity-stream==0.6.3
django-appconf==1.0.2
django-bootstrap-pagination==1.6.2
django-classy-tags==0.8.0
django-cms==3.4.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-filer==1.2.7
django-formtools==2.0
django-grappelli==2.6.5
django-haystack==2.4.1
django-jsonfield==0.9.19
django-mptt==0.8.7
django-picklefield==0.3.2
django-polymorphic==1.0.2
django-recaptcha==1.0.6
django-reversion==1.8.7
django-rosetta==0.7.13
django-sekizai==0.10.0
Django-Select2==5.10.0
django-sortedm2m==1.3.3
django-spurl==0.6.4
django-standard-form==1.1.1
django-tinymce==1.5.4
django-treebeard==4.1.1
django-uni-form==0.8.0
djangocms-admin-style==1.2.7
djangocms-column==1.7.0
djangocms-file==0.2.1
djangocms-flash==0.2.0
djangocms-googlemap==0.5.2
djangocms-inherit==0.2.2
djangocms-installer==0.7.9
djangocms-link==1.6.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.8.1
djangocms-video==0.2.0
easy-thumbnails==1.5
facebook-sdk==0.3.2
future==0.16.0
guess-language==0.2
hashids==1.2.0
html5lib==0.999999999
huey==1.3.1
idna==2.5
lxml==3.7.3
mailsnake==1.6.4
memcache-toolbar==0.5.6
microsofttranslator==0.8
MySQL-python==1.2.5
newrelic==2.86.3.70
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pipdeptree==0.10.1
polib==1.0.8
poster==0.8.1
psycopg2==2.7.1
pycryptodomex==3.4.6
pyDes==2.0.1
pyExcelerator==0.6.4.1
pyjwkest==1.3.2
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.2
raven==6.1.0
recaptcha-client==1.0.6
redis==2.10.5
requests==2.17.3
requests-oauthlib==0.8.0
selenium==2.46.1
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
sqlparse==0.2.3
twython==3.4.0
tzlocal==1.4
Unidecode==0.4.20
urllib3==1.21.1
URLObject==2.4.3
vobject==0.9.4.1
webencodings==0.5.1
Whoosh==2.5.7
YURL==0.13
